Reorganize impls for Display and From<AddrParseError> for NetAddrError
