+++ src/netaddr_error/from/addr_parse_error.rs
@@ -1,0 +1,8 @@
+use super::NetAddrError;
+use std::net::AddrParseError;
+
+impl From<AddrParseError> for NetAddrError {
+	fn from(other: AddrParseError) -> Self {
+		Self::ParseError(other.to_string())
+	}
+}
